fix run op
