+++ xdsl/interpreters/__init__.py
@@ -19,7 +19,6 @@ from xdsl.interpreters import (
     scf,
     snitch_stream,
     tensor,
-    test,
 )
 from xdsl.interpreters.experimental import pdl
 from xdsl.ir import MLContext
@@ -61,4 +60,3 @@ def register_implementations(
     interpreter.register_implementations(scf.ScfFunctions())
     interpreter.register_implementations(snitch_stream.SnitchStreamFunctions())
     interpreter.register_implementations(tensor.TensorFunctions())
-    interpreter.register_implementations(test.TestFunctions())
